게시물 수정

+++ src/main/java/com/ll/exam/article/repository/ArticleRepository.java
@@ -52,7 +52,7 @@ public class ArticleRepository {
         return sql.insert();
     }
 
-    public void modify(int id, String title, String body, boolean isBlind) {
+    public void modify(long id, String title, String body, boolean isBlind) {
         SecSql sql = myMap.genSecSql();
         sql
                 .append("UPDATE article")
